fix(typecheck): preserve strict member diagnostics
