+++ crates/vize_canon/src/virtual_ts/strict_template_globals_tests.rs
@@ -1,0 +1,126 @@
+use super::helpers::generate_template_context;
+use super::{TemplateGlobal, VirtualTsOptions, generate_virtual_ts_with_offsets};
+use vize_carton::config::VueVersion;
+use vize_croquis::{Analyzer, AnalyzerOptions};
+
+#[test]
+fn test_strict_template_context_keeps_router_but_not_unknown_plugin_globals() {
+    let ctx = generate_template_context(
+        &VirtualTsOptions {
+            strict_instance_globals: true,
+            ..Default::default()
+        },
+        VueVersion::V3,
+        false,
+    );
+
+    assert!(ctx.contains("type __VizeStrictTemplateContext"));
+    assert!(ctx.contains("$route: any;"));
+    assert!(ctx.contains("$router: any;"));
+    assert!(!ctx.contains("$t:"));
+}
+
+#[test]
+fn test_strict_template_unknown_refs_read_context_without_shadowing_auto_imports() {
+    let template = r#"<div>{{ getPreferences() }} {{ getPreferences() }} {{ currentUser.name }} {{ $shout('ok') }}</div>"#;
+
+    let allocator = vize_carton::Bump::new();
+    let (root, _) = vize_armature::parse(&allocator, template);
+
+    let mut analyzer = Analyzer::with_options(AnalyzerOptions::full());
+    analyzer.analyze_template(&root);
+    let summary = analyzer.finish();
+
+    let output = generate_virtual_ts_with_offsets(
+        &summary,
+        None,
+        Some(&root),
+        0,
+        0,
+        &VirtualTsOptions {
+            auto_import_bindings: vec!["currentUser".into()],
+            strict_instance_globals: true,
+            template_globals: vec![TemplateGlobal {
+                name: "$shout".into(),
+                type_annotation: "(value: string) => string".into(),
+                default_value: "((value: string) => value) as any".into(),
+            }],
+            ..Default::default()
+        },
+    );
+
+    assert!(
+        output
+            .code
+            .contains("var getPreferences: any = undefined; void (getPreferences);"),
+        "{}",
+        output.code
+    );
+    assert_eq!(
+        output
+            .code
+            .matches("__vize_strict_template_context.getPreferences")
+            .count(),
+        2,
+        "{}",
+        output.code
+    );
+    assert!(
+        output
+            .code
+            .contains("var currentUser: __U<__R_currentUser> = undefined as any;"),
+        "{}",
+        output.code
+    );
+    assert!(
+        !output
+            .code
+            .contains("__vize_strict_template_context.currentUser"),
+        "{}",
+        output.code
+    );
+    assert!(
+        !output
+            .code
+            .contains("__vize_strict_template_context.$shout"),
+        "{}",
+        output.code
+    );
+}
+
+#[test]
+fn test_strict_template_expression_keeps_member_root_type() {
+    let template = r#"<div>{{ route.missing }}</div>"#;
+    let script = "const route = { query: { tab: 'home' } };\n";
+
+    let allocator = vize_carton::Bump::new();
+    let (root, _) = vize_armature::parse(&allocator, template);
+
+    let mut analyzer = Analyzer::with_options(AnalyzerOptions::full());
+    analyzer.analyze_script_setup(script);
+    analyzer.analyze_template(&root);
+    let summary = analyzer.finish();
+
+    let output = generate_virtual_ts_with_offsets(
+        &summary,
+        Some(script),
+        Some(&root),
+        0,
+        0,
+        &VirtualTsOptions {
+            strict_instance_globals: true,
+            ..Default::default()
+        },
+    );
+
+    assert!(
+        !output.code.contains("var route: any = undefined;"),
+        "{}",
+        output.code
+    );
+    assert!(
+        output.code.contains("void (route.missing);"),
+        "{}",
+        output.code
+    );
+}
